+++ source/kernel/spawn.zig
@@ -42,7 +42,7 @@ pub fn spawn(allocator: std.mem.Allocator, entry: anytype, arg: ?*const anyopaqu
     }
 }
 
-pub fn root_process(entry: anytype, arg: ?*const anyopaque, stack_size: usize) !void {
+pub fn root_process(entry: anytype, arg: ?*const anyopaque, stack_size: u32) !void {
     try process_manager.instance.create_process(stack_size, entry, arg, "/");
     if (process_manager.instance.scheduler.schedule_next()) {
         process_manager.instance.initialize_context_switching();
